nghttp: Support SSLKEYLOGFILE
